requirements.txt
# ==========================================
# Core AI Clients
# ==========================================
openai>=1.12.0          # GPT-4o mini, Attack
google-generativeai>=0.4.0 # Gemini API

# ==========================================
# Data Processing & Analysis
# ==========================================
pandas>=2.1.0           # 데이터 로드 및 CSV/JSON 핸들링
numpy>=1.26.0           # 수치 연산 (Pandas 의존성)

# ==========================================
# Utilities
# ==========================================
tqdm>=4.66.0            # 진행률 바 (Progress Bar)
python-dotenv>=1.0.0    # .env 파일에서 API Key 보안 로드

# ==========================================
# Development & Formatting (Optional)
# ==========================================
black                   # 코드 포맷팅 (PEP8 준수)
isort                   # import 순서 자동 정렬
jupyter                 # 데이터 분석용 노트북